Allow uploadir to upload as streamable documents. TODO: need to make this configurable

+++ stdplugins/purge.py
@@ -6,21 +6,19 @@ import asyncio
 async def _(event):
     if event.fwd_from:
         return
-    i = 1
-    msgs = []
-    shiiinabot = "\u2060"
-    for i in range(4000):
-        shiiinabot += "\u2060"
-    await event.edit(shiiinabot)
-    await event.delete()
-    async for message in borg.iter_messages(event.chat_id, min_id=event.reply_to_msg_id, from_user="me"):
-        i = i + 1
-        try:
-            await message.edit(shiiinabot + message.message)
-        except:
-            pass
-        msgs.append(message)
-        await asyncio.sleep(3)
-    if len(msgs) <= 100:
-        await borg.delete_messages(event.chat_id, msgs)
+    if event.reply_to_msg_id:
+        i = 1
         msgs = []
+        shiiinabot = "\u2060"
+        for i in range(4000):
+            shiiinabot += "\u2060"
+        await event.edit(shiiinabot)
+        async for message in borg.iter_messages(event.chat_id, min_id=event.reply_to_msg_id):
+            i = i + 1
+            msgs.append(message)
+        if len(msgs) <= 100:
+            await borg.delete_messages(event.chat_id, msgs)
+            msgs = []
+            await event.delete()
+        else:
+            await event.edit("**PURGE** Failed!")
