+++ stdplugins/purgeme.py
@@ -7,16 +7,20 @@ async def _(event):
     if event.fwd_from:
         return
     i = 1
+    msgs = []
     shiiinabot = "\u2060"
-    for i in range(4006):
+    for i in range(4000):
         shiiinabot += "\u2060"
     await event.edit(shiiinabot)
-    await event.delete()
     async for message in borg.iter_messages(event.chat_id, from_user="me"):
         i = i + 1
-        try:
-            #  await message.edit(shiiinabot)
-            await message.delete()
-        except:
-            pass
-        await asyncio.sleep(10)
+        msgs.append(message)
+        if len(msgs) == 100:
+            await borg.delete_messages(event.chat_id, msgs)
+            msgs = []
+    if len(msgs) <= 100:
+        await borg.delete_messages(event.chat_id, msgs)
+        msgs = []
+        await event.delete()
+    else:
+        await event.edit("**PURGE** Failed!")
